Move package layout to simplified form
Only one package "main" needed (for now) for this relatively simple
program.  If a substantive case for packaging part of the codebase,
after it has evolved a bit, we can do so on an as-needed basis.  This
original packaging layout was partly just me learning the packaging, and
partly a result of adapting to and learning the Cobra automated
package/command setup.