Rename Namepair to Pathsplit
Mainly to indicate the content is specifically path related.  Also
abstract away some of the details by using the name "Path" or "Paths"
for variables, rather than the specific representation type.

Currently the splitting of the pathname has little benefit, but down the
line it is expected to pool or intern the path and filename components
in order to save memory (especially for the dirnames which are
especially redundant with lots of different filenames)