Get LINK_MAX Pathconf value from getconf command
The Go lang syscall (or x/sys) system exposes Pathconf() to the BSD
unixes, but not Linux (apparently), so I can't just call
Pathconf(pathname, PC_LINK_MAX) to get the maximum number of hard links
allowed for a specific filesystem device.  As a workaround, I call the
getconf system command (which is part of glibc-common on Linux, and the
system_cmds subsystem on MacOs; I haven't checked all the other BSDs but
OpenBSD has it installed by default, so it seems reasonable that getconf
will be available without fail on Unix-like systems).

Since the getconf command is only called once per file-system device, it
will have a negligable performance impact.  The _POSIX_LINK_MAX
definition, which is the minimum supported value for LINK_MAX, is an
amazingly low value of 8.  I use that value as a fallback for now, but
may choose to raise it later (it's only a fallback if the getconf data
retrieval fails, which shouldn't happen anyway).