Add function to append a reversed Ino slice
The appending of reversed slices, that are later undone on a subsequent
loops, will mainly allow the equal filename only matching work (when
implemented), although it also works here to help with handling the
maximum nlinks being reached in the source inode.  The
appendReversedInos() helper function simplifies the implementation of
these special cases.