Move srcPath lookup in sortedLinks to inside loop
This involves repeatedly looking up the src pathname inside the
innermost loop, the value of which will come from implementing the
equal-filenames option, in which it'll be desirable to search for a
matching filename for the source inode for a given destination filename.

This will allow us to efficiently output linkable pairs that always have
matching filenames, when so desired.  Otherwise it may have gotten the
same result, but by referring to an inode by it's non-matching filename,
which would be confusing.