Add contentDigest function for classifying content
Allow computing a hash (ie. digest) of the first part of a file.  This
can be used later to quickly determine files which cannot possibly match
another file (because its digest is different), and reduce the search
space for equal files.  The stored digests need not be cryptographically
secure, just have reasonably good collision avoidance behavior.

Identical files are still always compared byte-by-byte, so the digests
main purpose is to exclude files that are definitely not identical,
rather than deciding whether they are identical.  Only the beginning of
files are checked, because the file comparison checks perform that
initial I/O anyway, so this approach is meant to not add an appreciable
I/O burden to the algorithm (as the file comparison I/O is the slowest
part of the algorithm).