Allow using LinearSearchThresh to enable digests
If LinearSearchThresh is non-negative, it can determine the maximum
length of the cachedInodeSet to is reached before switching over to
usage of content digests to exclude definitely unequal files.  As the
length of the cachedInodeSets increase (such as when encountering many
files with equal lengths but unequal content), the use of digests can
greatly reduce the linear searches, which is a potentially O(N**2)
search as N (the number of matching inode hashes) increases.