Add useDigest bool to determine when to use digest
In order to use digests, the LinearSearchThresh option has to be
non-negative, and the cachedIno set has to exceed the thresh value.
Store this condition into a boolean to pass to the
areFilesHardlinkable() method, which helps properly track the statistics
on how many digests are computed, and not compute digests unnecessarily.