Remove comparisons from progress output
This is to help keep the progress output short, and the comparisons
value was mostly useful if/when the linear-search-threshold command
option was changed, but the comparison information can be seen at the
program's end.  It may make sense to print it when the search thresh is
changed from the default value.