Add xattr comparison function
Compare the xattrs of two files, returning true if they are equal (same
keys and values).  Relies on third part xattr functions, which are path
based.  Ideally we would be able to open the file and use the fd
versions of the system xattr calls, but those don't appear to be
supported on all the unixes.