Output bytes saved for each kind of inode mismatch
When an equal file is found, but there are some inode parameters that
don't match (because they are ignored), keeping track of the bytes that
wouldn't have been saved if the parameter wasn't ignored can be useful
for deciding whether to change the options or perform the linking.  This
change outputs the bytes associated with each individual parameter.