Add filenamePaths type for consistent link pairs
Remember the pathname returned when an arbitrary pathname is requested,
so that it can be returned on subsequent requests.  This helps ensure
the same src path is used to link with multiple destinations.  Its not
really required for correctness, but benefits the user by providing link
pair results which coalesce when output.