Use pathsplitSet to reduce pathname lookup to O(1)
The previous use of slices allowed pathnames to have a consistent order,
helping to keep arbitrary pathname lookups somewhat consistent.
However, it didn't apply when multiple filenames were involved. Now that
we store the arbitrary pathname directly, we can switch to an unordered
set that can make find a pathname in a long list of pathnames more
efficient.