Switch InoPaths map to use pointer values
The filenamePaths struct type uses some internal caching, which means
it's methods modify the struct.  This means the modified values have to
be reassinged to the InoPaths map when using value semantics.

By switching to using pointer values for the InoPaths map, the actual
structs get modified (they are not help in the map, only pointers to
them) and the amount of code to maintain is reduced.