Add simple string interning pool
Basically used equality checking (via map storage) to find an equal
string in the pool and reuse it.  The idea is that the underlying string
storage will be reused (ie. multiple string object with the same string
backing data); the cost is the lookup for equal strings (and the garbage
collection for the discarded string object)