Use timer goroutine to rate-limit progress updates
Use a bidirectional channel to indicate when enough time has passed to
update the progress output.  This helps reduce the amount of printing
when progress is very fast (so that we aren't printing for every single
file), while also ensuring we update at a reasonable speed when the rate
of progress is low.