Return linkedInoSets in "sorted" order
Rather than building the linkedInoSets in random order (from map key
iteration), we first sort the keys, which are Ino numbers.  This
produces more consistent orderings when run multiple times without
changes to the underlying file system.  The end result is that the
returned link pairs are more likely to be in the same order (or close to
it), which is a usability improvement.