Calculate file/dir count stats from unique paths
The previous method of counting while we walked the tree, could leave to
overcounts when subdirectories were specified multiple times on the
command line.  Even with simple checking for repeated dir and file
strings, specifying a subdirectory would subvert this check.  The new
method should ensure an exact count of unique dir and file paths.