Begin removal of MyCLIOptions global
Move rootCmd initialization into init function, and setup CLIOptions
variable beforehand to pass to cliRun() func and to use in flags
initialization.  By doing so, and passing the CLIOptions on to the
runnable functions, we don't need it as a global anymore.  This was a
vestige of the cobra framework setup code, which placed the rootCmd
setup outside of init().