Detect changes to on-disk Nlink value when linking
Like other file changes, we can detect if the Nlink value has changed
when actual linking is enabled; this could include when a link has been
created from outside the walked tree, but in any case is another way of
detecting that the walked filesystem is not in a quiescent state.  It
requires that the on-disk file stat updates to match what is being
recorded in the hardlinkable StatInfo structs (ie. our model of the file
system stat structures must be kept in sync when we update the
filesystem).