Remove SortedLinks() goroutine, used for chan send
The SortedLinks goroutine was used to send hardlinkable pairs to
receivers, which as a side effect generated the finished link Results
from the nlink-sorted and linked inodes.  However, the goroutine made
passing errors from the hardlinkFiles func harder to propogate.  Since
the SortedLinks public method wasn't currently being used for anything
other than it's side effect, the goroutine was removed and the methods
renamed (to generateLinks() and generateLinksHelper()) to better
indicate their purpose.  The channels used to pass link pairs back from
the goroutine were also removed.

It's possible to get the LinkPaths information from the Results struct
anyway (which also indicates whether the Result generation was aborted
before completion), so using a channel to return linkable pairs is not
necessary.