Move linkedInoSets() funcs to internal/inode pkg
These were methods of the fsDev type, even though they operated solely
on the LinkedInos member of the type.  So, move them to the inode pkg to
reflect their limited scope, and also to ensure that the linkedInoSets()
goroutine (renamed to All()) cannot race with other methods of the fsDev
type.

The new Add() method and All() goroutine can (hypothetically) race, but
won't be used simultaneously so no locking is needed.