Create inoDigests type with receiver methods
Create a new type to consolidate the Digest operations, and move the
fsDev methods to use the new type as a receiver.

Keeping the Results updated, now that the digest methods no longer have
access to the fsDev status members, requires checking the digest
creation error status for each usage of the digest creation functions.
The benefit of removing a dependency on an unrelated type is worth the
small amount of extra responsibility.