Move digest accounting out of contentDigest()
The digest computed Result can no longer use the fsDev receiver to
update, so do the accounting with return values.

An error value is not enough to accurately count the amount of digests
computed; it can cause overcounts by ignoring when an existing digest
for a given ino is known to exist.  Instead return a bool from newDigest
to specifically indicate when a digest was computed.  We ignore errors
because missing digests don't change the outcome of the algorithm, just
possibly lead to more file comparisons attempted.