Iterate over a slice of pathnames in AllPaths()
As it turns out, while iterating over the PathsMap and modifying it
would have worked for a single-threaded algorithm, by having the map
reading iterations in a different goroutine from the one modifying it
can lead to panic().  Rather than changing to sync.Map(), we instead
resort to making a new slice of all the pathnames to iterate over inside
the goroutine.  The slice creation should be somewhat faster than doing
a full copy of the nested PathMap, and safe to iterate over while the
original PathMap is updated.  The additional memory usage is modest
(since the pathname objects themselves need not be duplicated).