Rename LinkedInoSets to LinkableInoSets
LinkedInoSets gave the impression that it (perhaps) represented already
linked inodes, which is not really a valid concept, but could sound
similar to 'linked files', which are really pathnames that are linked to
a single inode.

So instead, to better represent that we are finding inodes that could be
"linked" together (by moving all pathnames from one inode to the other,
and eliminating one of the inodes altogether), use the term "linkable"
instead of "linked".  Ie. these are inodes that are not linked, but
found to be "linkable".