Allow accurate new link prediction using clusters
With the addition of existing links to the randomized testing, we need
to track which pathnames are linked together before the Run(), and
compare with the results.  These pre-linked pathnames are called
"clusters", and we predict the outcome by sorting from the largest to
smallest clusters, and assume that paths representing equivalent file
content will all become linked to the largest cluster.