Remove foundDirectory() call from walk and delete
While this func can keep a running tally of encountered directories
during the walk, the results are eventually overwritten when all the
pathnames are counted, and the exact number of directories is counted.
This is because the same directories could be specified multiple times
by the user, and this func would overcount them.

Since we don't need the directory running count for the progress
interfaces (like we do the running file count), remove the
foundDirectory() call from the walk, and remove the now unused func.