Add check of NewLinkCount for SameName option
The "clustering" of links affects this algorithm, because the "same
filename" requirement doesn't ensure that inodes which are hardlinkable
(ie. have equal file content, etc) will actually all be linked into
fewer inodes.  So where a pathname ends up depends on the sorting of the
clusters, etc.

This test does a basic check that the algorithm gives equal results of
what would be expected, just by counting the number of filenames
encountered in the sorted "cluster" list (ie. without manipulating the
cluster list itself).