Disable min/max rnd filesize in samename rand test
The ignored files can still be linked (on-disk) to the other files,
which can change the sort order of the on-disk clusters as compared to
the test cluster list.  This can cause the NewLinkCount value to
occasionally not match test value, usually just by 1 or 2.

To avoid this, we disable min/max file size filtering for the SameName
randomized test.  In principle we could account for the filesize
filtered filenames in the test clusters, but that's complex enough that
for now we avoid it (maybe it'd be better as a separate test).