Allow CheckQuiescence option for non-linking runs
Allow checking to see if the filesystem is being modified (by previously
scanned files having changed inodes) even when linking is not enabled.
The cost is an extra stat for each linkable file, which generally isn't
too much of a penalty, especially if it's important that the returned
run results be stable.