Change cli --linking-enabled to --enable-linking
Turn the CLI option into a verb, to be consistent with other CLI
options.  Options is still adjective form (LinkingEnabled) to better
match "if" statement usage.

Also indicate that --enable-linking option sets the --quiescence option
(to abort if the walked files are found to be modified during or after
the filesystem walk).