Use time.Time for StatInfo for portability
Stat_t types have different field names for time values on different OS;
even just amongst the Unix types.  Darwin used Mtimespec, for example,
whereas Linux was Mtim.  Even the BSDs sometimes use Mtim and differ
from Darwin.  This is true for the Atimespec and Ctimespec fields as
well. It's kind of a mess.

Luckily we only need the Mtimespec field, and the os.Lstat call has a
ModTime() method that portably returns the modification time as a
time.Time type.