Add UseNewestLink Option to set link time/uid/gid
After moving a link from one inode to another (in dolink), this Option
determines whether to perform an update of the src inode to the
mtime/atime and uid/gid of the newer dst inode (which is the default).
The motivation is so that file paths that are linked will appear to have
their modification times updated to a newer time, rather than going back
in time.  For backup-programs this will ensure that modified links get
backed up, rather than overlooked.