Proper handling of cmp reads less than buffer size
At the end of the file, less than the total amount of buffer bytes will
be read into the file comparison buffers.  Reslice to the size of the
read bytes, to ensure we are only comparing the most recently read
bytes.

Without this change, the algorithm would continue to work, because the
previous bytes were equal, it just meant more bytes were compared than
necessary.  However because the buffers are now reused, and not
initialized on each use, its safer to ensure we are only ever comparing
the most recently read bytes (and slightly more efficient).
Furthermore, prevent buffer slice from being accidentally enlarged at
the end of file reads, using 'atEnd' bool.