Move dir and file validation to runHelper
Move the CLI validation to the runHelper, directory and file validation
(ie. removing duplicates, and ensuring the passed args are only dirs or
regular files).  Also remove the redundant min/max file size option
validation out of CLIRun.

Rather than requiring dirs and files to be provided in different slices,
have Run accept just a single slice, and split them into dirs and files
in runHelper().  This makes it easier on Run() users, and removes the
need for redundant validations (in CLIRun and Run).