Move StringPool into status struct
Still accessible to fsDev, but also makes it easier to use in
non-fsDev functions.  But rather than a pool per fsDev, this makes one
pool for all.  This should be okay; multiple fsDevs are rare, and it
allows the non-fsDev users to benefit from a pool of directory paths
accross multiple devices.