Add panic guard if walk gives a non-regular file
There is a bug in godirwalk before v1.7.4 that could allow a non-regular
file (such as a symlink) to be passed back from the walk; when we Lstat
the file, double check that it is regular and abort if not.

Bug reported to godirwalk author (karrick) and fixed for next release.