Make file compare use *os.File instead of Reader
Just to make it clear we are comparing files specifically, and not
io.Reader interfaces in general.  In particular, we don't make an effort
to support a partial Read(), except in cases of reaching EOF.  By
defining the file w/ *os.File args, we don't imply it can support any
io.Reader().